Story/ncl 2370 commit arm64 specific changes for civetweb (#7)
* Added arm64 config and Update openssl api version for debug-x86 config

* Resolved data loss warnings in ARM64 configuration build

+++ src/civetweb.c
@@ -22520,7 +22520,7 @@ mg_getclient_socket(const struct mg_connection * conn)
 {
 	if(conn)
 	{
-		return conn->client.sock;
+		return (unsigned int) conn->client.sock;
 	}
 	else
 	{
